feature/clientes
agrego los cambios del toast y un par de arreglos mas

+++ controladores/ClienteCtr.php
@@ -15,6 +15,12 @@ class ClienteCtr
         $action = isset($_GET['action']) ? $_GET['action'] : '';
         $module = isset($_GET['module']) ? $_GET['module'] : '';
         $id = isset($_GET['id']) ? $_GET['id'] : '';
+        $status = isset($_GET['status']) ? $_GET['status'] : "";
+        $toast = new ToastCtr();
+        if ($status == "error") {
+            $description = isset($_GET['description']) ? $_GET['description'] : "";
+            $toast->mostrarToast($status, $description);
+        }
         if ($module == 'clientes') {
             switch ($action) {
                 case 'created':
@@ -23,13 +29,30 @@ class ClienteCtr
                         if ($status != "success") {
                             $this->create();
                         }
+                    } else {
+                        if ($status == "success") {
+                            $toast->mostrarToast("exito", "Cliente creado");
+                        }
                     }
                     break;
                 case 'deleted':
-                    $this->delete($id);
+                    if ($status != "success") {
+                        $this->delete($id);
+                    } else if ($status == "success") {
+                        $toast->mostrarToast("exito", "Cliente eliminado");
+                    }
                     break;
                 case 'edited':
-                    $this->update($id);
+                    if ($_SERVER["REQUEST_METHOD"] == "POST") {
+                        $status = isset($_GET['status']) ? $_GET['status'] : "";
+                        if ($status != "success") {
+                            $this->update($id);
+                        }
+                    } else {
+                        if ($status == "success") {
+                            $toast->mostrarToast("exito", "Cliente editado");
+                        }
+                    }
                     break;
                 case 'searched':
                     $this->search();
@@ -97,7 +120,12 @@ class ClienteCtr
 
     public function getAllClientes()
     {
-        return $this->clienteDAO->getAllClientes();
+        $clientes = $this->clienteDAO->getAllClientes();
+        if (is_string($clientes)) {
+            $toast = new ToastCtr();
+            $toast->mostrarToast("error", "error al traer los clientes", $clientes);
+        }
+        return $clientes;
     }
 
     public function getPantallaDelete()
@@ -109,7 +137,12 @@ class ClienteCtr
 
     public function getClienteById($id)
     {
-        return $this->clienteDAO->getClienteById($id);
+        $cliente = $this->clienteDAO->getClienteById($id);
+        if (is_string($cliente)) {
+            $toast = new ToastCtr();
+            $toast->mostrarToast("error", "error al traer el cliente", $cliente);
+        }
+        return $cliente;
     }
 
     public function search()
